+++ src/components/pages/common/wallets-list/index.tsx
@@ -17,9 +17,10 @@ import ZerionWalletIcon from 'images/zerion-wallet.png'
 import RainbowWalletIcon from 'images/rainbow-wallet.png'
 import ImtokenWalletIcon from 'images/imtoken-wallet.png'
 import WalletConnectIcon from 'images/walletconnect-wallet.png'
+import CrossmintIcon from 'images/crossmint-wallet.png'
 import ENSIcon from 'images/ens-logo.png'
 import { useConnect, Connector } from 'wagmi'
-import { TDropStep, TMultiscanStep, TWalletName, TWalletOption } from 'types'
+import { TDropStep, TMultiscanStep, TWalletName, TWalletOption, TDropType } from 'types'
 import { AdditionalNoteComponent } from 'linkdrop-ui'
 import {  OverlayScreen } from 'linkdrop-ui'
 import * as dropAsyncActions from 'data/store/reducers/drop/async-actions'
@@ -78,9 +79,12 @@ const isOptionVisible = (
   option: TWalletOption | undefined,
   preferredWallet: string | null,
   currentOption: string,
-  availableWallets: string[]
+  availableWallets: string[],
+  additionalCondition?: boolean
 ) => {
-  
+  if (additionalCondition !== undefined && !additionalCondition) {
+    return undefined
+  }
   if (!option) { return undefined }
   if (!availableWallets || availableWallets.length === 0 || currentOption === preferredWallet) {
     return option
@@ -91,11 +95,12 @@ const isOptionVisible = (
 }
 
 const defineOptionsList = (
+  type: TDropType | null,
   setStep: (step: TDropStep & TMultiscanStep) => void,
   open: (options?: any | undefined) => Promise<void>,
   connect: (args: Partial<any> | undefined) => void,
   connectors: Connector<any, any, any>[],
-  wallet: string | null,
+  wallet: TWalletName | null,
   deeplinkRedirect: (
     deeplink: string,
     walletId: TWalletName
@@ -108,7 +113,7 @@ const defineOptionsList = (
 
   const system = defineSystem()
   const ensOption = !isManual && enableENS ? {
-    title: 'Enter ENS or address',
+    title: 'ENS or address',
     onClick: () => setStep('set_address'),
     icon: <WalletIcon src={ENSIcon} />
   } : undefined
@@ -121,6 +126,16 @@ const defineOptionsList = (
     icon: <WalletIcon src={WalletConnectIcon} />,
     recommended: wallet === 'walletconnect'
   }
+
+  const crossmintOption = {
+    title: 'Sign in with email',
+    onClick: () => {
+      setStep('crossmint_connection')
+    },
+    icon: <WalletIcon src={CrossmintIcon} />,
+    recommended: wallet === 'crossmint'
+  }
+
   const injected = connectors.find(connector => connector.id === "injected")
   const injectedOption = getInjectedWalletOption(
     wallet,
@@ -147,9 +162,10 @@ const defineOptionsList = (
 
     const wallets = [
       isOptionVisible(injectedOption, wallet, 'metamask', availableWallets),
+      isOptionVisible(crossmintOption, wallet, 'crossmint', availableWallets, type !== 'ERC20'),
       isOptionVisible(coinbaseOption, wallet, 'coinbase_wallet', availableWallets),
       isOptionVisible(walletConnectOption, wallet, 'walletconnect', availableWallets),
-      ensOption
+      isOptionVisible(ensOption, wallet, 'manual_address', availableWallets)
     ]
 
     return sortWallets(wallets) 
@@ -228,7 +244,8 @@ const defineOptionsList = (
     isOptionVisible(coinbaseOption, wallet, 'coinbase_wallet', availableWallets),
     isOptionVisible(zerionOption, wallet, 'zerion', availableWallets),
     isOptionVisible(walletConnectOption, wallet, 'walletconnect', availableWallets),
-    ensOption,
+    isOptionVisible(crossmintOption, wallet, 'crossmint', availableWallets, type !== 'ERC20'),
+    isOptionVisible(ensOption, wallet, 'manual_address', availableWallets),
     isOptionVisible(imtokenOption, wallet, 'imtoken', availableWallets),
     isOptionVisible(trustOption, wallet, 'trust', availableWallets),
     isOptionVisible(rainbowOption, wallet, 'rainbow', availableWallets)
@@ -245,7 +262,8 @@ const WalletsList: FC<ReduxType> = ({
   campaignId,
   deeplinkRedirect,
   availableWallets,
-  enableENS
+  enableENS,
+  type
 }) => {
   const { open } = useWeb3Modal()
   const { connect, connectors } = useConnect()
@@ -254,6 +272,7 @@ const WalletsList: FC<ReduxType> = ({
   const injected = connectors.find(connector => connector.id === "injected")
 
   const options = defineOptionsList(
+    type,
     setStep,
     open,
     connect,
@@ -265,6 +284,7 @@ const WalletsList: FC<ReduxType> = ({
     availableWallets,
     enableENS
   )
+
 
   return <Container>
     <TitleComponent>Connect your wallet</TitleComponent>