+++ src/configs/wallets.tsx
@@ -14,11 +14,11 @@ type TWallet = {
   mobile: {
     android: {
       support: boolean,
-      deepLink: TWalletDeeplink
+      deepLink?: TWalletDeeplink
     },
     ios: {
       support: boolean,
-      deepLink: TWalletDeeplink
+      deepLink?: TWalletDeeplink
     }
   }
 }
@@ -29,7 +29,7 @@ const wallets: TWallets = {
   metamask: {
     id: 'metamask',
     name: 'MetaMask',
-    chains: [1, 5, 137, 80001],
+    chains: [1, 5, 137, 80001, 8453, 84531],
     mobile: {
       android: {
         support: true,
@@ -38,6 +38,19 @@ const wallets: TWallets = {
       ios: {
         support: true,
         deepLink: (url: string) => `https://metamask.app.link/dapp/${_withoutProtocol(url)}`
+      }
+    },
+  },
+  crossmint: {
+    id: 'crossmint',
+    name: 'Crossmint',
+    chains: [1, 5, 137, 80001, 8453, 84531],
+    mobile: {
+      android: {
+        support: false
+      },
+      ios: {
+        support: false,
       }
     },
   },
@@ -56,7 +69,7 @@ const wallets: TWallets = {
         deepLink: (url: string) => `https://get.status.im/browse/${_withoutProtocol(url)}`
       }
     },
-    chains: [1, 5, 137, 80001]
+    chains: [1, 5, 137, 80001, 8453, 84531]
   },
   trust: {
     id: 'trust',
@@ -73,7 +86,7 @@ const wallets: TWallets = {
         deepLink: (url: string) => `https://link.trustwallet.com/open_url?coin_id=60&url=${encodeURIComponent(url)}`
       }
     },
-    chains: [1, 5, 137, 80001]
+    chains: [1, 5, 137, 80001, 8453, 84531]
   },
   opera: {
     id: 'opera',
@@ -91,7 +104,7 @@ const wallets: TWallets = {
         deepLink: (url: string) => `https://operacrypto.page.link/?link=${encodeURIComponent(url)}&efr=1&ibi=com.opera.OperaTouch&isi=1411869974`
       }
     },
-    chains: [1, 5, 137, 80001]
+    chains: [1, 5, 137, 80001, 8453, 84531]
   },
   coinbase_wallet: {
     id: 'coinbase_wallet',
@@ -108,7 +121,7 @@ const wallets: TWallets = {
         deepLink: (url: string) => `https://go.cb-w.com/dapp?cb_url=${encodeURIComponent(url)}`
       }
     },
-    chains: [1, 5, 137, 80001]
+    chains: [1, 5, 137, 80001, 8453, 84531]
   },
   imtoken: {
     id: 'imtoken',
@@ -127,7 +140,7 @@ const wallets: TWallets = {
           `imtokenv2://navigate/DappView?url=${encodeURIComponent(url)}`
       }
     },
-    chains: [1, 5, 137, 80001]
+    chains: [1, 5, 137, 80001, 8453, 84531]
   },
   zerion: {
     id: 'zerion',
@@ -144,7 +157,7 @@ const wallets: TWallets = {
         deepLink: (url: string) => null
       }
     },
-    chains: [1, 5, 137, 80001]
+    chains: [1, 5, 137, 80001, 8453, 84531]
   },
   rainbow: {
     id: 'rainbow',
@@ -162,7 +175,33 @@ const wallets: TWallets = {
       }
     },
     chains: [1]
-  }
+  },
+  walletconnect: {
+    id: 'walletconnect',
+    name: 'WalletConnect',
+    chains: [1, 5, 137, 80001, 8453, 84531],
+    mobile: {
+      android: {
+        support: false
+      },
+      ios: {
+        support: false,
+      }
+    },
+  },
+  manual_address: {
+    id: 'manual_address',
+    name: 'ENS or address',
+    chains: [1, 5, 137, 80001, 8453, 84531],
+    mobile: {
+      android: {
+        support: false
+      },
+      ios: {
+        support: false,
+      }
+    },
+  },
 }
 
 export default wallets