+++ src/data/store/reducers/drop/async-actions/get-link-by-multi-qr.tsx
@@ -4,10 +4,12 @@ import { DropActions } from '../types'
 import { ethers } from 'ethers'
 import * as actionsDrop from '../actions'
 import { plausibleApi, getMultiQRData } from 'data/api'
-import { checkIfMultiscanIsPresented } from 'helpers'
 import axios, { AxiosError } from 'axios'
-import { RootState } from 'data/store'
 import * as wccrypto from '@walletconnect/utils/dist/esm'
+import { RootState } from 'data/store'
+import { defineJSONRpcUrl } from 'helpers'
+import * as asyncActionsDrop from './'
+const { REACT_APP_INFURA_ID } = process.env
 
 export default function getLinkByMultiQR(
   multiscanQRId: string,
@@ -18,12 +20,23 @@ export default function getLinkByMultiQR(
   callback: (location: string) => void
 ) {
   return async (
-    dispatch: Dispatch<DropActions>
+    dispatch: Dispatch<DropActions>,
+    getState: () => RootState
   ) => {
     dispatch(actionsDrop.setLoading(true))
     dispatch(actionsDrop.setError(null))
 
-    try {
+    const {
+      drop: {
+        previewSetting,
+        tokenAddress,
+        type,
+        tokenId,
+        chainId
+      }
+    } = getState()
+
+    try {      
       const { data } = await getMultiQRData(
         multiscanQRId,
         scanId,
@@ -32,14 +45,32 @@ export default function getLinkByMultiQR(
       )
 
       const { encrypted_claim_link, success }: { encrypted_claim_link: string, success: boolean } = data
-      
       if (success && encrypted_claim_link) {
         const decryptKey = ethers.utils.id(multiscanQREncCode)
         const linkDecrypted = wccrypto.decrypt({ encoded: encrypted_claim_link, symKey: decryptKey.replace('0x', '') })
         if (linkDecrypted && callback) {
           callback(linkDecrypted)
         }
-      } 
+      }
+
+
+      if (previewSetting === 'token') {
+        const jsonRpcUrl = defineJSONRpcUrl({ chainId: Number(chainId), infuraPk: REACT_APP_INFURA_ID as string })
+        const provider = new ethers.providers.JsonRpcProvider(jsonRpcUrl)
+
+        if (
+          type && tokenAddress && chainId
+        ) {
+          await asyncActionsDrop.getTokenData(
+            type,
+            tokenAddress,
+            tokenId,
+            chainId,
+            provider,
+            dispatch
+          )
+        }
+      }
 
 
       dispatch(actionsDrop.setLoading(false))