+++ src/helpers/define-json-rpc-url.tsx
@@ -3,13 +3,15 @@ const {
   REACT_APP_JSON_RPC_POLYGON,
   REACT_APP_JSON_RPC_MAINNET,
   REACT_APP_JSON_RPC_GOERLI,
-  REACT_APP_JSON_RPC_MUMBAI
+  REACT_APP_JSON_RPC_MUMBAI,
+  REACT_APP_JSON_RPC_BASE,
+  REACT_APP_JSON_RPC_BASEGOERLI
 } = process.env
 
 const defineJSONRpcUrl = ({ chainId, infuraPk } : { chainId: number, infuraPk: string }) => {
   const networkName = defineNetworkName(chainId)
 
-  if (networkName === 'matic') {
+  if (networkName === 'polygon') {
     return REACT_APP_JSON_RPC_POLYGON
   } else if (networkName === 'mumbai') {
     return REACT_APP_JSON_RPC_MUMBAI
@@ -17,6 +19,10 @@ const defineJSONRpcUrl = ({ chainId, infuraPk } : { chainId: number, infuraPk: s
     return REACT_APP_JSON_RPC_GOERLI
   } else if (networkName === 'mainnet') {
     return REACT_APP_JSON_RPC_MAINNET
+  } else if (networkName === 'base') {
+    return REACT_APP_JSON_RPC_BASE
+  } else if (networkName === 'baseGoerli') {
+    return REACT_APP_JSON_RPC_BASEGOERLI
   }
   
   return `https://${networkName}.infura.io/v3/${infuraPk}`