+++ src/types/multiscan-step.tsx
@@ -3,6 +3,7 @@ type TMultiscanStep = 'initial' |
                  'download_await' |
                  'zerion_connection' |
                  'wallet_redirect_await' |
-                 'set_address'
+                 'set_address' |
+                 'crossmint_connection'
 
 export default TMultiscanStep