+++ src/types/retro-drop.tsx
@@ -1,9 +1,0 @@
-type TRetroDrop = {
-  title: string,
-  ipfsHash: string,
-  address: string,
-  chainId: number,
-  description: string
-}
-
-export default TRetroDrop