+++ src/types/wallet-name.tsx
@@ -1,3 +1,13 @@
-type TWalletName = 'metamask' | 'status' | 'trust' | 'opera' | 'coinbase_wallet' | 'imtoken' | 'zerion' | 'rainbow'
+type TWalletName = 'metamask' |
+                   'status' |
+                   'trust' |
+                   'opera' |
+                   'coinbase_wallet' |
+                   'imtoken' |
+                   'zerion' |
+                   'rainbow' |
+                   'crossmint' |
+                   'walletconnect' |
+                   'manual_address'
 
 export default TWalletName