+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Linkdrop Claim app
 
+## 2.2.14
+- added support for dispenser redirect
+- fix for custom redirect button on "claiming finished" page
+- Crossmint option added to wallets
+- updated logic of connection to wallets
+- added support for Base and BaseGoerli networks
+- added fix for Mumbai network explorer url
+
 ## 2.2.13
 - added event for "Powered by Linkdrop" click
 - added "Powered by Linkdrop" on "already_claimed" and "claiming_finished" screens